+++ client/src/extension.ts
@@ -1,16 +1,28 @@
 import * as path from "path";
-import { workspace, ExtensionContext } from "vscode";
+import { languages, workspace, ExtensionContext } from "vscode";
 
 import {
+  DocumentFilter,
   LanguageClient,
   LanguageClientOptions,
   ServerOptions,
   TransportKind
 } from "vscode-languageclient/node";
+import { provideKickstartCompletionItems } from "./completion";
+
+/**
+ * Document selector for files that this extension can enhance.
+ */
+const DOCUMENT_SELECTOR: DocumentFilter = { scheme: "file", language: "kickstart" };
 
 let client: LanguageClient;
 
 export function activate(context: ExtensionContext) {
+  const completionProvider = languages.registerCompletionItemProvider(DOCUMENT_SELECTOR, {
+    provideCompletionItems: provideKickstartCompletionItems
+  });
+  context.subscriptions.push(completionProvider);
+
   const serverModule = context.asAbsolutePath(
     path.join("out", "server", "src", "server.js")
   );
@@ -24,7 +36,7 @@ export function activate(context: ExtensionContext) {
   };
 
   const clientOptions: LanguageClientOptions = {
-    documentSelector: [{ scheme: "file", language: "kickstart" }],
+    documentSelector: [DOCUMENT_SELECTOR],
     synchronize: {
       fileEvents: workspace.createFileSystemWatcher("**/*.ks")
     }